Refactor on phone formatting and controller
